Added `is_signed` to `mgpu.FragmentedArray`
The registers within a fragmented array always use signless types, and instead
the signedness is tracked on the fragmented arrays itself (i.e. in Python).

Note that the Mosaic GPU lowering in Pallas does not yet specify signedness.
I will do that in a follow up to avoid inflating the diff here.

PiperOrigin-RevId: 676417875